refactor(nuxt-ripple): reshuffled mapping utils to be less confusingly named
