+++ packages/nuxt-ripple/mapping/utils/formatPriceRange.test.ts
