+++ packages/nuxt-ripple/mapping/utils/formatPriceRange.ts
