+++ packages/ripple-tide-grant/mapping/utils/index.ts
@@ -1,0 +1,1 @@
+export { default as formatGrantAudiences } from './formatGrantAudiences.js'
