--- packages/ripple-tide-grant/utilities/index.ts
@@ -1,1 +1,0 @@
-export { default as formatGrantAudiences } from '../utilities/formatGrantAudiences.js'
