+++ packages/ripple-tide-grant/utils/index.ts
@@ -1,1 +1,1 @@
-export { default as formatGrantAudiences } from '../utilities/formatGrantAudiences'
+export { default as formatGrantAudiences } from '../mapping/utils/formatGrantAudiences'
